Extract structure panel into 2 components: NormalStructurePanel and FloatingStructurePanel for clarity
